[DONE] property owner deleting bug

+++ server/src/app/modules/propertyOwner/propertyOwner.service.ts
@@ -444,6 +444,7 @@ const deletePropertyOwnerData = async (propertyOwnerId: string): Promise<any | n
       },
     });
 
+    const userId = propertyOwner?.userId;
     if (!propertyOwner) {
       throw new ApiError(httpStatus.NOT_FOUND, "Property Owner Not Found!!!");
     }
@@ -538,14 +539,18 @@ const deletePropertyOwnerData = async (propertyOwnerId: string): Promise<any | n
     }
 
     // ! removing all properties owned by property owner
-    const removingAllProperties = await transactionClient.property.deleteMany({
-      where: {
-        ownerId: propertyOwnerId,
-      },
-    });
-
-    if (!removingAllProperties) {
-      throw new ApiError(httpStatus.BAD_REQUEST, "Properties removing Failed");
+    const getCountOfProperty = await transactionClient.property.count();
+    console.log("count", getCountOfProperty);
+    if (getCountOfProperty > 0) {
+      const removingAllProperties = await transactionClient.property.deleteMany({
+        where: {
+          ownerId: propertyOwnerId,
+        },
+      });
+      console.log("count deleting", removingAllProperties);
+      if (!removingAllProperties?.count) {
+        throw new ApiError(httpStatus.BAD_REQUEST, "Properties removing Failed");
+      }
     }
     // ! removing property owner
     const removingPropertyOwner = await transactionClient.propertyOwner.delete({
@@ -557,18 +562,26 @@ const deletePropertyOwnerData = async (propertyOwnerId: string): Promise<any | n
     if (!removingPropertyOwner) {
       throw new ApiError(httpStatus.BAD_REQUEST, "Property Owner removing Failed");
     }
-    // ! removing property owner user data
-    const removingPropertyOwnerUser = await transactionClient.user.delete({
-      where: {
-        userId: propertyOwner?.userId,
-      },
-    });
 
-    if (!removingPropertyOwnerUser) {
-      throw new ApiError(httpStatus.BAD_REQUEST, "Property Owner User removing Failed");
+    console.log("shafin1", userId);
+
+    // ! removing property owner user data
+    try {
+      const removingPropertyOwnerUser = await transactionClient.user.delete({
+        where: {
+          userId: userId,
+        },
+      });
+      if (!removingPropertyOwnerUser) {
+        throw new ApiError(httpStatus.BAD_REQUEST, "Property Owner User removing Failed");
+      }
+      console.log("ahamori", removingPropertyOwnerUser);
+    } catch (error) {
+      console.log("error", error);
     }
 
     //
+    // return;
     return removingPropertyOwner;
   });
